Migrate ContextDetectorTest heredoc patterns to fixtures
Create non-autoloaded fixtures for incomplete code patterns:
- function_body_incomplete.php
- multiline_comment_open.php
- docblock_open.php
- heredoc_open.php
- nowdoc_open.php
- heredoc_closed.php
