+++ tests/Fixtures/ContextDetector/docblock_open.php
@@ -1,0 +1,3 @@
+<?php
+/**
+ * @param string $var
