+++ tests/Fixtures/ContextDetector/function_body_incomplete.php
@@ -1,0 +1,4 @@
+<?php
+function test(): void
+{
+    $x = new
