+++ tests/Fixtures/ContextDetector/heredoc_closed.php
@@ -1,0 +1,5 @@
+<?php
+$x = <<<HTML
+content
+HTML;
+$this->
