+++ tests/Fixtures/ContextDetector/heredoc_open.php
@@ -1,0 +1,3 @@
+<?php
+$x = <<<HTML
+<div>Some content
