+++ tests/Fixtures/ContextDetector/multiline_comment_open.php
@@ -1,0 +1,3 @@
+<?php
+/*
+ * Comment $this->
