+++ tests/Fixtures/ContextDetector/nowdoc_open.php
@@ -1,0 +1,3 @@
+<?php
+$x = <<<'TEXT'
+Some text content
